BaseLocaleListener
