+++ src/libs/Localize/index.desktop.js
@@ -1,0 +1,3 @@
+import ELECTRON_EVENTS from "../../../desktop/ELECTRON_EVENTS";
+
+window.electron.send(ELECTRON_EVENTS.LOCALE_UPDATED, val);
