+++ src/libs/OptionsListUtils.js
@@ -6,7 +6,7 @@ import Str from 'expensify-common/lib/str';
 import ONYXKEYS from '../ONYXKEYS';
 import CONST from '../CONST';
 import * as ReportUtils from './ReportUtils';
-import * as Localize from './Localize';
+import * as Localize from './Localize/Localize';
 import Permissions from './Permissions';
 import * as CollectionUtils from './CollectionUtils';
 import Navigation from './Navigation/Navigation';
