+++ src/libs/actions/BankAccounts.js
@@ -2,7 +2,7 @@ import Onyx from 'react-native-onyx';
 import CONST from '../../CONST';
 import * as API from '../API';
 import ONYXKEYS from '../../ONYXKEYS';
-import * as Localize from '../Localize';
+import * as Localize from '../Localize/Localize';
 import DateUtils from '../DateUtils';
 
 export {
