+++ src/libs/actions/Report.js
@@ -20,7 +20,7 @@ import Log from '../Log';
 import * as LoginUtils from '../LoginUtils';
 import * as ReportUtils from '../ReportUtils';
 import Growl from '../Growl';
-import * as Localize from '../Localize';
+import * as Localize from '../Localize/Localize';
 import DateUtils from '../DateUtils';
 import * as ReportActionsUtils from '../ReportActionsUtils';
 import * as OptionsListUtils from '../OptionsListUtils';
